Publish into wiki repo

Publish an specific rmarkdown into wiki repository, specifying the path of the wiki repository once that you already clone this repository into your machine. See Details if is your first time publishing in a wiki repository.

publish_wiki(
  rmarkdown,
  path_wiki_repo,
  automatic_update = FALSE,
  quiet_render_logs = FALSE
)

Arguments

rmarkdown

the path of rmarkdown file

path_wiki_repo

the complete path of wiki repository.

automatic_update

if you like to do automatically the pull, commit and push set in TRUE, if you like to do manually, set FALSE. Default is TRUE.

quiet_render_logs

if you like to see the render logs of rmarkdown, set in TRUE, in otherwise set FALSE. Default is TRUE.

Value

Publish a md into wiki repository

Details

Publish an rmarkdown into wiki repo

If this is the first time you are going to publish in a wiki, you must follow these steps:

The wiki repository must be cloned to a local working folder. To do this first go to the wiki tab of your repository and you will need to manually create the first wiki page. If you don't do this it will tell you that the repository does not exist.

Once you have created the wiki home page, you will need to clone the wiki repository just like you would any other repository on your computer (using the link that is made available to you once you create the wiki home page).

Once you have cloned the wiki on your computer, what you should indicate in the publica_wiki function is the path_wiki_repo. This is the full address on your computer where this repository you cloned is located. For example: /home/client/project/wiki_repo

Examples

if (FALSE) {
publish_wiki(rmarkdown = "eda.Rmd",
             path_wiki_repo = "/home/cliente/proyecto/wiki_repo")
}